refactor(suite-common): share useStakingEntryPeriodEstimateInDays hook
