+++ suite-common/staking/src/index.ts
@@ -1,4 +1,5 @@
 export * from './utils';
 export type * from './types';
 export * from './actions';
+export * from './hooks';
 export * from './constants/ethereumNetworkAddresses';
